relocate getMostReadUrls
